make ts add-p&l.ts from work

+++ frontend/src/components/expenses.js
@@ -40,17 +40,28 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Expenses = void 0;
-var custom_http_ts_1 = require("../services/custom-http.ts");
+var custom_http_1 = require("../services/custom-http");
 var config_1 = __importDefault(require("../../config/config"));
-var show_categories_ts_1 = require("../services/show-categories.ts");
+var show_categories_1 = require("../services/show-categories");
 var sidebar_1 = require("./sidebar");
+var bootstrap_1 = __importDefault(require("bootstrap"));
 var Expenses = /** @class */ (function () {
     function Expenses() {
         this.editCategoryButtons = null;
         this.deleteCategoryButtons = null;
+        this.addCategoryButton = document.getElementById('add-category');
+        if (this.addCategoryButton) {
+            this.addCategoryButton.onclick = function () {
+                location.href = "#/add-expenses";
+            };
+        }
         //определяем параметры модальных окон
-        this.resultModal = new bootstrap.Modal(document.getElementById('textModal'));
-        this.confirmationModal = new bootstrap.Modal(document.getElementById('confirmationModal'));
+        var textModalElement = document.getElementById('textModal');
+        var confirmationModalElement = document.getElementById('confirmationModal');
+        if (textModalElement && confirmationModalElement) {
+            this.resultModal = new bootstrap_1.default.Modal(textModalElement);
+            this.confirmationModal = new bootstrap_1.default.Modal(confirmationModalElement);
+        }
         this.modalMessageField = document.getElementById('textModal-message');
         this.textMessage = null;
         this.getCategories();
@@ -66,16 +77,16 @@ var Expenses = /** @class */ (function () {
                         _a.label = 2;
                     case 2:
                         _a.trys.push([2, 9, , 10]);
-                        return [4 /*yield*/, custom_http_ts_1.CustomHttp.request(config_1.default.host + '/categories/expense', 'GET')];
+                        return [4 /*yield*/, custom_http_1.CustomHttp.request(config_1.default.host + '/categories/expense', 'GET')];
                     case 3:
                         result = _a.sent();
                         if (!result) return [3 /*break*/, 8];
                         if (!(result.error || !result)) return [3 /*break*/, 5];
-                        return [4 /*yield*/, this.showResult(result.message)];
+                        return [4 /*yield*/, this.showResult(result)];
                     case 4:
                         _a.sent();
                         throw new Error(result.message);
-                    case 5: return [4 /*yield*/, show_categories_ts_1.ShowCategories.init(result)];
+                    case 5: return [4 /*yield*/, show_categories_1.ShowCategories.init(result)];
                     case 6:
                         _a.sent(); //отрисуем карточки категорий
                         return [4 /*yield*/, this.processCategories()];
@@ -95,15 +106,11 @@ var Expenses = /** @class */ (function () {
     ;
     Expenses.prototype.processCategories = function () {
         return __awaiter(this, void 0, void 0, function () {
-            var _loop_1, _i, _a, element;
+            var deleteButtonsArray, _loop_1, _i, deleteButtonsArray_1, element;
             var _this = this;
-            return __generator(this, function (_b) {
+            return __generator(this, function (_a) {
                 this.editCategoryButtons = document.querySelectorAll('[id^="edit-"]');
                 this.deleteCategoryButtons = document.querySelectorAll('[id^="delete-"]');
-                this.addCategoryButton = document.getElementById('add-category');
-                this.addCategoryButton.onclick = function () {
-                    location.href = "#/add-expenses";
-                };
                 this.editCategoryButtons.forEach(function (element) {
                     element.addEventListener("click", function () {
                         var id = element.id;
@@ -111,6 +118,7 @@ var Expenses = /** @class */ (function () {
                         location.href = '#/edit-expenses?=' + number;
                     });
                 });
+                deleteButtonsArray = Array.from(this.deleteCategoryButtons);
                 _loop_1 = function (element) {
                     element.addEventListener("click", function () { return __awaiter(_this, void 0, void 0, function () {
                         var id, number;
@@ -130,8 +138,8 @@ var Expenses = /** @class */ (function () {
                         });
                     }); });
                 };
-                for (_i = 0, _a = this.deleteCategoryButtons; _i < _a.length; _i++) {
-                    element = _a[_i];
+                for (_i = 0, deleteButtonsArray_1 = deleteButtonsArray; _i < deleteButtonsArray_1.length; _i++) {
+                    element = deleteButtonsArray_1[_i];
                     _loop_1(element);
                 }
                 return [2 /*return*/];
@@ -146,25 +154,31 @@ var Expenses = /** @class */ (function () {
                         var deleteButton = document.getElementById('delete');
                         var cancelButton = document.getElementById('cancel');
                         var confirmationText = document.getElementById('confirmationText');
-                        confirmationText.innerText = 'Вы действительно хотите удалить категорию?';
+                        if (confirmationText) {
+                            confirmationText.innerText = 'Вы действительно хотите удалить категорию?';
+                        }
                         _this.confirmationModal.show();
-                        cancelButton.onclick = function () {
-                            _this.confirmationModal.hide();
-                            resolve(); // Разрешаем обещание после закрытия модального окна
-                        };
-                        deleteButton.onclick = function () { return __awaiter(_this, void 0, void 0, function () {
-                            return __generator(this, function (_a) {
-                                switch (_a.label) {
-                                    case 0:
-                                        this.confirmationModal.hide();
-                                        return [4 /*yield*/, this.deleteCategory(categoryId)];
-                                    case 1:
-                                        _a.sent();
-                                        resolve(); // Разрешаем обещание после удаления категории
-                                        return [2 /*return*/];
-                                }
-                            });
-                        }); };
+                        if (cancelButton) {
+                            cancelButton.onclick = function () {
+                                _this.confirmationModal.hide();
+                                resolve(); // Разрешаем обещание после закрытия модального окна
+                            };
+                        }
+                        if (deleteButton) {
+                            deleteButton.onclick = function () { return __awaiter(_this, void 0, void 0, function () {
+                                return __generator(this, function (_a) {
+                                    switch (_a.label) {
+                                        case 0:
+                                            this.confirmationModal.hide();
+                                            return [4 /*yield*/, this.deleteCategory(categoryId)];
+                                        case 1:
+                                            _a.sent();
+                                            resolve(); // Разрешаем обещание после удаления категории
+                                            return [2 /*return*/];
+                                    }
+                                });
+                            }); };
+                        }
                     })];
             });
         });
@@ -175,27 +189,29 @@ var Expenses = /** @class */ (function () {
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
-                        if (!categoryId) return [3 /*break*/, 6];
+                        if (!categoryId) return [3 /*break*/, 8];
                         _a.label = 1;
                     case 1:
-                        _a.trys.push([1, 5, , 6]);
-                        return [4 /*yield*/, custom_http_ts_1.CustomHttp.request(config_1.default.host + '/categories/expense/' + categoryId, 'DELETE')];
+                        _a.trys.push([1, 7, , 8]);
+                        return [4 /*yield*/, custom_http_1.CustomHttp.request(config_1.default.host + '/categories/expense/' + categoryId, 'DELETE')];
                     case 2:
                         result = _a.sent();
-                        if (!result) return [3 /*break*/, 4];
-                        if (result.error || !result) {
-                            throw new Error();
-                        }
+                        if (!result) return [3 /*break*/, 6];
+                        if (!(result.error || !result)) return [3 /*break*/, 4];
                         return [4 /*yield*/, this.showResult(result)];
                     case 3:
                         _a.sent();
-                        _a.label = 4;
-                    case 4: return [3 /*break*/, 6];
+                        throw new Error();
+                    case 4: return [4 /*yield*/, this.showResult(result)];
                     case 5:
+                        _a.sent();
+                        _a.label = 6;
+                    case 6: return [3 /*break*/, 8];
+                    case 7:
                         error_2 = _a.sent();
                         console.log('ошибка' + error_2);
-                        return [3 /*break*/, 6];
-                    case 6: return [2 /*return*/];
+                        return [3 /*break*/, 8];
+                    case 8: return [2 /*return*/];
                 }
             });
         });
@@ -207,7 +223,9 @@ var Expenses = /** @class */ (function () {
                 return [2 /*return*/, new Promise(function (resolve) {
                         _this.textMessage = message.error ? message.message :
                             "Категория успешно удалена." + "\nСообщение сервера: " + JSON.stringify(message);
-                        _this.modalMessageField.innerText = _this.textMessage;
+                        if (_this.modalMessageField) {
+                            _this.modalMessageField.innerText = _this.textMessage;
+                        }
                         _this.resultModal.show();
                         // Обработчик события при закрытии попапа
                         _this.resultModal._element.addEventListener('hidden.bs.modal', function () {
